bussiness logic moved from repository to services and refactor

+++ src/services/studentService.js
@@ -1,8 +1,14 @@
 const studentRepository = require("../repositories/studentRepository");
+const teacherRepository = require("../repositories/teacherRepository");
 
 const getAllStudents = async () => {
   try {
     const allStudents = await studentRepository.getAllStudents();
+
+    if (!allStudents || allStudents.length === 0) {
+      throw { status: 404, message: "No students found" };
+    }
+
     return allStudents;
   } catch (error) {
     throw error;
@@ -12,6 +18,11 @@ const getAllStudents = async () => {
 const getStudentById = async (studentId) => {
   try {
     const studentById = await studentRepository.getStudentById(studentId);
+
+    if (!studentById) {
+      throw { status: 404, message: "Student not found for provided id" };
+    }
+
     return studentById;
   } catch (error) {
     throw error;
@@ -20,6 +31,26 @@ const getStudentById = async (studentId) => {
 
 const createNewStudent = async (studentData) => {
   try {
+    const studentByDni = await studentRepository.getStudentByDni(studentData);
+
+    if (studentByDni) {
+      throw {
+        status: 400,
+        message: "Student already registered with provided dni",
+      };
+    }
+
+    const teacherExists = await teacherRepository.getTeacherById(
+      studentData.TeacherId
+    );
+
+    if (!teacherExists) {
+      throw {
+        status: 404,
+        message: "Teacher not found with provided teacher id",
+      };
+    }
+
     const newStudent = await studentRepository.createNewStudent(studentData);
     return newStudent;
   } catch (error) {
@@ -29,6 +60,18 @@ const createNewStudent = async (studentData) => {
 
 const updateStudent = async (studentId, studentData) => {
   try {
+    const isAlreadyRegistered = await studentRepository.getStudentById(
+      studentId
+    );
+
+    if (!isAlreadyRegistered) {
+      throw { status: 404, message: "Student not found with provided id" };
+    }
+
+    if (isAlreadyRegistered.TeacherId !== studentData.TeacherId) {
+      throw { status: 403, message: "Cannot update teacher id" };
+    }
+
     const updatedStudent = await studentRepository.updateStudent(
       studentId,
       studentData
@@ -41,6 +84,14 @@ const updateStudent = async (studentId, studentData) => {
 
 const deleteStudent = async (studentId) => {
   try {
+    const isAlreadyRegistered = await studentRepository.getStudentById(
+      studentId
+    );
+
+    if (!isAlreadyRegistered) {
+      throw { status: 404, message: "Student not found with provided id" };
+    }
+
     const deletedStudent = await studentRepository.deleteStudent(studentId);
     return deletedStudent;
   } catch (error) {
