+++ src/services/userService.js
@@ -1,8 +1,12 @@
 const userRepository = require("../repositories/userRepository");
+const teacherRepository = require("../repositories/teacherRepository")
 
 const getAllUsers = async () => {
   try {
     const allUsers = await userRepository.getAllUsers();
+    if (!allUsers || allUsers.length === 0) {
+      throw { status: 404, message: "No users found" };
+    }
     return allUsers;
   } catch (error) {
     throw error;
@@ -12,6 +16,11 @@ const getAllUsers = async () => {
 const getUserById = async (userId) => {
   try {
     const userById = await userRepository.getUserById(userId);
+
+    if (!userById) {
+      throw { status: 404, message: "User not found for provided user id" };
+    }
+
     return userById;
   } catch (error) {
     throw error;
@@ -20,6 +29,15 @@ const getUserById = async (userId) => {
 
 const createNewUser = async (userData) => {
   try {
+    const isAlreadyRegistered = await userRepository.getUserByEmail(userData.email);
+
+    if (isAlreadyRegistered) {
+      throw {
+        status: 400,
+        message: "User already registered with provided email",
+      };
+    }
+
     const newUser = await userRepository.createNewUser(userData);
     return newUser;
   } catch (error) {
@@ -29,6 +47,11 @@ const createNewUser = async (userData) => {
 
 const updateUser = async (userId, userData) => {
   try {
+    const userById = await userRepository.getUserById(userId);
+    if (!userById) {
+      throw { status: 404, message: "User not found provided id" };
+    }
+
     const updatedUser = await userRepository.updateUser(userId, userData);
     return updatedUser;
   } catch (error) {
@@ -38,6 +61,21 @@ const updateUser = async (userId, userData) => {
 
 const deleteUser = async (userId) => {
   try {
+    const user = await userRepository.getUserById(userId);
+    console.log(user);
+
+    if (!user) {
+      throw { status: 404, message: "User not found for provided user id" };
+    }
+
+    const hasTeacher = await teacherRepository.getTeacherByUserId(userId);
+    if (hasTeacher) {
+      throw {
+        status: 403,
+        message: "User cannot be deleted because it has teacher associated",
+      };
+    }
+
     const deletedUser = await userRepository.deleteUser(userId);
     return deletedUser;
   } catch (error) {
@@ -47,8 +85,23 @@ const deleteUser = async (userId) => {
 
 const checkAndUpdateUserStatus = async (userId) => {
   try {
-    const activeUser = await userRepository.checkAndUpdateUserStatus(userId);
-    return activeUser;
+    const userById = await userRepository.getUserById(userId);
+    
+    if (!userById) {
+      throw { status: 404, message: "User not found provided id" };
+    }
+
+    if (userById.active === true) {
+      return userById;
+    } else {
+      const userData = {
+        email: userById.email,
+        password: userById.password,
+        active: true
+      }
+      const activeUser  = await userRepository.updateUser(userId, userData);
+      return activeUser
+    }
   } catch (error) {
     throw error;
   }
@@ -56,8 +109,12 @@ const checkAndUpdateUserStatus = async (userId) => {
 
 const checkUserStatus = async (userId) => {
   try {
-    const userStatus = await userRepository.checkUserStatus(userId);
-    return userStatus;
+    const userById = await userRepository.getUserById(userId);
+    if (!userById) {
+      throw { status: 404, message: "User not found provided id"};
+    }
+    const userStatus = userById.active;
+    return userStatus
   } catch (error) {
     throw error;
   }
